requirements.txt
click==8.1.3
Flask==2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
SQLAlchemy==2.0.7
typing_extensions==4.5.0
Werkzeug==2.2.3
python-dotenv==1.0.0
requests==2.28.2